Move state management deeper to RoleMenuGroupsSection component
